Replace float64 with float32 #23

+++ tf2rl/algos/categorical_dqn.py
@@ -50,7 +50,7 @@ class CategoricalDQN(DQN):
         self._delta_z = (self._v_max - self._v_min) / (self.q_func._n_atoms - 1)
         self._z_list = tf.constant(
             [self._v_min + i * self._delta_z for i in range(self.q_func._n_atoms)],
-            dtype=tf.float64)
+            dtype=tf.float32)
         self._z_list_broadcasted = tf.tile(
             tf.reshape(self._z_list, [1, self.q_func._n_atoms]),
             tf.constant([self._action_dim, 1]))
@@ -63,7 +63,7 @@ class CategoricalDQN(DQN):
         if not test and np.random.rand() < self.epsilon:
             action = np.random.randint(self._action_dim)
         else:
-            state = np.expand_dims(state, axis=0).astype(np.float64)
+            state = np.expand_dims(state, axis=0).astype(np.float32)
             action_probs = self._get_action_body(tf.constant(state))
             action = tf.argmax(
                 tf.reduce_sum(action_probs * self._z_list_broadcasted, axis=2),
@@ -96,7 +96,7 @@ class CategoricalDQN(DQN):
                 tf.reshape(done ,[-1, 1]),
                 tf.constant([1, self.q_func._n_atoms]))  # [batch_size, n_atoms]
             discounts = tf.cast(
-                tf.reshape(self.discount, [-1, 1]), tf.float64)
+                tf.reshape(self.discount, [-1, 1]), tf.float32)
             z = tf.reshape(
                 self._z_list, [1, self.q_func._n_atoms])  # [1, n_atoms]
             z = rewards + not_done * discounts * z  # [batch_size, n_atoms]
